fix: loading/saving persisted state, status race condition
Since Maps cannot be JSON-stringified when saved or JSON-parsed when read back, this adds a replacer and reviver function for serializing and deserializing (respectively) the `state.wallets` Map in persisted state.

This also fixes the race condition caused by directly mutating the state object in store mutation functions. Now state is updated immutably, and the mutation functions all return a full state object (even when an error occurs).