feat: resume connected wallet sessions with WalletManager.resumeSessions
Once the wallet manager and store are initialized and the DOM is ready, this public method can be used to call each connected wallet client's `resumeSession` method.

Only the WalletConnect clients and Exodus clients actually contain logic. Exodus checks for the existence of the extension and its `isConnected` property.

The Pera/Defly/Daffi and WalletConnect clients return their connected accounts from storage. This handles the edge case of those accounts not matching what was stored in persisted state, updating the store with the returned accounts.

In the future this method may be removed, as it's really only dealing with edge cases that could be handled during initialization.