feat: initialize wallet clients on demand
This refactors the WalletManager > Wallet > WalletClient pattern by combining the Wallet and WalletClient classes into a single class. Instead of having one class to interact with the wallet client itself and another for exposing the API and managing state, the BaseWallet class handes both.

Instead of initializing the wallet clients when the WalletManager is instantiated, they are initialized "on demand", only when the user first connects (or if a session is resumed). When dynamic imports are supported, this will mean that the massive client/SDK packages are not imported until they are actually needed.