feat: wallet metadata
This adds a static property defaultMetadata to the BaseWallet class, which specifies the wallet's name and base64 icon. These properties can be overwritten by defining a different `name` or `icon` (base64 string or file path) in the `metadata` property of any wallet's configuration object.

This also refactors some of the wallet type names, referring to wallets as "wallets" instead of "clients" so it's easier to understand.