refactor: move types/utils for supported wallets to same folder
The test runner was complaining about `createWalletMap` in src/wallets/utils.ts, which is a function that doesn't need to be tested.

After moving `createWalletMap` to its own file colocated with the supported wallets, it made sense to do the same for everything related to specific wallets.

Now, when support for a new wallet is added, only files inside the src/wallets/supported folder will need to be added or updated.

And now I can test the other functions in src/wallets/utils