refactor: convert class methods to regular functions
Automatic mocking and spying of class methods does not work if they are arrow functions. (The use of regular/arrow functions hasn't been particularly consistent either.)

Now all class methods are regular functions, with the exception of `subscribe` and `notifySubscribers` in WalletManager.

Since those methods are passed to wallet class constructors as callbacks, they must inherit `this` from the enclosing scope where they are defined: the WalletManager class. Otherwise, `this.subscribers` will be undefined in the wallet classes when they are called.