feat(core): convert state.wallets from a Map to an Object
Through experimentation it seems that the cause of reactivity bugs in the framework adapters was the result of `state.wallets` being stored as a Map. Maps are not immutable and changes were not triggering renders since the variable's reference wasn't changing.

This also simplifies (de)serialization of persistent storage, since there is no longer a need for the replacer/reviver functions to stringify/parse the JSON.