feat(react): use @tanstack/react-store, return reactive Wallet objects
This completes the transition to @tanstack/store by using its React adapter for reactive changes to the `WalletManager`'s store.

The `Wallet` objects in the exported `wallets` array are a combination of `BaseWallet` methods and derived state variables that will trigger renders in the UI when their values change.

Tests for `WalletProvider` are included, as well as initial tests for the `useWallet` hook (more are needed).